Add storage provider for requiring user enters a password to access
If the user doesn't have a cookie we first redirect them to the enter password page and they enter a password. We then assign them an Id and write down whether it was a day guest or evening guest. The Id will later be used as the "storer" so we know who they entered