Give each editable thing its own Subject (observable thing)
There was a weird feature where a previous update was affecting a future element, something to do with the change to the new thing not pushing? No idea

I assume there's a more concise way of doing this rather than all the duplicate code but I can't be bothered to work it out now.